Use ascii encoding to unpack message into buffer
signPersonalMessage decodes the message as ascii, but Buffer.from()
defaults to using a utf-8 decoder. For binary strings such as hashes
and unicode data, this can lead to inaccurate decoding. Using the ascii
decoder ensures that the binary buffer corresponds to the original
message.

No test is added with this commit, as the existing tests of this
function only ensure that a event listener is created.