Remove BIP44 accounts without indexes during deserialization
The index of each BIP44 account is saved when that account is connected
to ensure we can determine the account's HD path later on for signing.
This should work for any new accounts that are added, but pre-existing
ledger accounts won't have this index, so signing will fail.

Accounts with missing indexes are now removed when deserializing, so
that it's more obvious to the user that they need to re-add them.