Rename confusing internal function (#63)
The `_isBIP44` internal function has been renamed to
`_isLedgerLiveHdPath`, which better represents what it does. It checks
to see whether the current HD path matches the Ledger Live HD path,
which is distinct from the BIP44 standard HD path.

The keyring state still uses a `bip44` boolean flag on each
`accountDetails` entry to represent this. A TODO comment has been added
to rename that state at some point in the future, as it is also
misleading. That has not been done in this PR because it is more
complicated, because it needs a migration and might be relied upon by
consumers of this package.