Ensure `this.app` gets constructed if transport check succeeds (#82)
Previously `this.app` would only get constructed if the transport check
failed. This relied upon the assumption that the transport would not
work if it hadn't been established yet. However the transport check
can succeed the first time if using Ledger Live, if the bridge to
Ledger Live has already been established.

`this.app` is now always constructed in `makeApp` if it's falsy. The
`reestablish` boolean is used to ensure we don't needlessly construct
`this.app` every time `makeApp` is called - we only re-construct it if
the connection was dropped and had to be re-established.